+++ cdk/services/serverless/import-service/ImportService.ts
@@ -1,9 +1,13 @@
 import { Construct } from 'constructs';
-import { ImportBucket } from '../../s3/import/importBucket';
+//import { ImportBucket } from '../../s3/import/importBucket';
 import { ImportProductsFile } from './importProductsFile';
 import { ApiGatewayImportService } from './ApiGatewayImportService';
-import { S3 } from '../../../constants/constants';
+import { QUEUE, S3 } from '../../../constants/constants';
 import { ImportFileParser } from './importFileParser'; 
+import { taskNum } from '../../../constants/constants';
+import { Bucket, EventType } from 'aws-cdk-lib/aws-s3';
+import { LambdaDestination } from "aws-cdk-lib/aws-s3-notifications";
+//import { S3EventSource } from 'aws-cdk-lib/aws-lambda-event-sources';
 
 
 class ImportService extends Construct {
@@ -11,17 +15,50 @@ class ImportService extends Construct {
         super(scope, id);
 
         //const importBucket = new ImportBucket(scope, S3.IMPORT_BUCKET_ID).importBucket;
-        //const backetName = importBucket.bucketName; 
+        //const backetName = importBucket.bucketName;
+
+        const importBucket = Bucket.fromBucketName(
+            scope,
+            S3.IMPORT_BUCKET_ID,
+            S3.IMPORT_BUCKET_NAME,
+        );
 
         const backetName = S3.IMPORT_BUCKET_NAME; 
 
-        const importProductsFileLambda = new ImportProductsFile(scope, 'importProductsFileLambda', backetName).importProductsFileLambda
+        const importProductsFileLambda = new ImportProductsFile(scope, `importProductsFileLambda-${taskNum}`, backetName).importProductsFileLambda;
+
+        importBucket.grantReadWrite(importProductsFileLambda);
+        importBucket.grantPut(importProductsFileLambda);
  
-        const apiGatewayImportService = new ApiGatewayImportService( scope, 'apiGatewayImportService',
+        const apiGatewayImportService = new ApiGatewayImportService( scope, `apiGatewayImportService-${taskNum}`,
             importProductsFileLambda,
         ).apiGatewayProductService;
 
-        const importFileParserLambda = new ImportFileParser( scope, 'importFileParserLambda').importFileParserLambda;
+        const importFileParserLambda = new ImportFileParser( scope, `importFileParserLambda-${taskNum}`, backetName).importFileParserLambda;
+
+        importBucket.grantReadWrite(importFileParserLambda);
+        importBucket.grantPut(importFileParserLambda);
+        importBucket.grantDelete(importFileParserLambda);
+
+        importBucket.addEventNotification(
+            EventType.OBJECT_CREATED,
+            new LambdaDestination(importFileParserLambda),
+            {
+                prefix: S3.IMPORT_UPLOADED_PREFIX,
+                suffix: S3.IMPORT_UPLOADED_SUFFIX,
+            }
+        )
+/*
+        importFileParserLambda.addEventSource(
+            new S3EventSource(importBucket as Bucket, {
+              events: [EventType.OBJECT_CREATED],
+              filters: [
+                { prefix: S3.IMPORT_UPLOADED_PREFIX }, 
+                { suffix: S3.IMPORT_UPLOADED_SUFFIX},
+            ],
+            })
+        );
+*/
 
     }
 }
