+++ cdk/services/serverless/import-service/importFileParser.ts
@@ -1,27 +1,19 @@
 import { Construct } from 'constructs';
 import { IFunction } from 'aws-cdk-lib/aws-lambda';
 import * as lambda from 'aws-cdk-lib/aws-lambda';
-import { S3EventSource } from 'aws-cdk-lib/aws-lambda-event-sources'; 
-import { Bucket ,EventType } from 'aws-cdk-lib/aws-s3'; 
-import { S3 } from '../../../constants/constants';
+import { QUEUE, LOGGING, S3 } from '../../../constants/constants';
 import { RetentionDays } from 'aws-cdk-lib/aws-logs';
+import { Queue } from 'aws-cdk-lib/aws-sqs';
 
 
 class ImportFileParser {
     importFileParserLambda: IFunction
     
-
     constructor(scope: Construct, id: string,
-        //importBucketName: string, 
+        importBucketName: string, 
         ) {
 
-        const importBucket = Bucket.fromBucketName(
-            scope,
-            S3.IMPORT_BUCKET_ID,
-            S3.IMPORT_BUCKET_NAME
-        );
-        //const importBucketArn = `arn:aws:s3:::${S3.IMPORT_BUCKET_NAME}`;
-        //const importBucket = Bucket.fromBucketArn(scope, S3.IMPORT_BUCKET_ID, importBucketArn);
+        const catalogItemsQueue = Queue.fromQueueArn(scope, QUEUE.CATALOG_ITEMS_QUEUE_ID, QUEUE.CATALOG_ITEMS_QUEUE_ARN);
 
         // Define the Lambda function resource
         const importFileParserLambda = new lambda.Function(scope, id, {
@@ -29,22 +21,15 @@ class ImportFileParser {
             code: lambda.Code.fromAsset('dist/import-service/lambdas'), // Points to the lambda directory
             handler: 'importFileParserLambda.handler', // Points to the 'importProductsFileLambda' file in the lambda directory
             environment: {
-                IMPORT_BUCKET_NAME: importBucket.bucketName,
-                LOG_LEVEL: 'INFO',
+                IMPORT_BUCKET_NAME: importBucketName,
+                IMPORT_UPLOADED_PREFIX: S3.IMPORT_UPLOADED_PREFIX,
+                IMPORT_SQS_URL: catalogItemsQueue.queueUrl,
+                LOG_LEVEL: LOGGING.IMPORT_FILE_PARSER,
             },
             logRetention: RetentionDays.ONE_WEEK,
         });
 
-        importBucket.grantReadWrite(importFileParserLambda);
-        importBucket.grantPut(importFileParserLambda);
-        importBucket.grantDelete(importFileParserLambda);
-
-        importFileParserLambda.addEventSource(
-            new S3EventSource(importBucket as Bucket, {
-              events: [EventType.OBJECT_CREATED],
-              filters: [{ prefix: 'uploaded/' }],
-            })
-          );
+        catalogItemsQueue.grantSendMessages(importFileParserLambda);
 
         this.importFileParserLambda = importFileParserLambda;
