+++ package.json
@@ -27,6 +27,8 @@
     "typescript": "^5.4.5"
   },
   "dependencies": {
+    "@aws-sdk/client-sns": "^3.609.0",
+    "@aws-sdk/client-sqs": "^3.609.0",
     "@aws-sdk/lib-storage": "^3.600.0"
   }
 }
